+++ Sources/Commander/ArgumentDescription.swift
@@ -350,7 +350,8 @@ class Help : Error, ANSIConvertible, CustomStringConvertible {
     if let group = group {
       output.append("Commands:")
       output.append("")
-      for command in group.commands {
+      let commands = group.commands.sorted(by: { $0.name < $1.name })
+      for command in commands {
         if let description = command.description {
           output.append("    + \(command.name) - \(description)")
         } else {
@@ -412,7 +413,8 @@ class Help : Error, ANSIConvertible, CustomStringConvertible {
     if let group = group {
       output.append("Commands:")
       output.append("")
-      for command in group.commands {
+      let commands = group.commands.sorted(by: { $0.name < $1.name })
+      for command in commands {
         if let description = command.description {
           output.append("    + \(ANSI.green)\(command.name)\(ANSI.reset) - \(description)")
         } else {