+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Bug Fixes
 
+
+- Only print errors in red if the output terminal supports ANSI color codes.
+  [#58](https://github.com/kylef/Commander/pull/58)
+
 - Showing default values for custom `ArgumentConvertible` types are now
   supported in the `--help` output of commands.