+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Commander Changelog
 
-## Master
+## 0.9.0 (2019-06-12)
 
 ### Breaking